requirements.txt
click==7.1.2
cycler==0.10.0
decorator==4.4.2
dominate==2.6.0
Flask==1.1.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.3
flaskr==1.0.0
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.3
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
opencv-contrib-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.17.0
PyWavelets==1.1.1
scikit-image==0.18.1
scipy==1.6.2
six==1.15.0
tifffile==2021.4.8
visitor==0.1.3
Werkzeug==1.0.1
WTForms==2.3.3
